+++ CHANGELOG.md
@@ -5,6 +5,12 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Serialization/deserialization using serde.
+
 ## [0.3.0] - 2020-06-05
 
 ### Added
@@ -34,6 +40,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Principal component analysis (PCA).
 
+[0.4.0]: https://github.com/petabi/petal-decomposition/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/petabi/petal-decomposition/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/petabi/petal-decomposition/compare/0.1.1...0.2.0
 [0.1.1]: https://github.com/petabi/petal-decomposition/compare/0.1.0...0.1.1